requirements.txt
pip==23.0.1
setuptools==65.5.1
wheel==0.38.4
pytube~=12.1.2
discord.py[voice]==2.1.0
pydantic~=1.10.4
pydantic[dotenv]
fastapi==0.95.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
starlette==0.26.1
python-multipart==0.0.6
pydantic[email]
uvicorn[standard]==0.21.1
SQLAlchemy==1.4.41
asyncpg==0.27.0
sqlmodel==0.0.8
alembic==1.10.4
httpx==0.24.0
requests==2.29.0